grsec: Apply grsecurity-2.9.1-2.6.32.59-201209192117.patch
commit 2353e62e599b143357c375d4844c25e3da8cefc2
Author: Brad Spengler <spender@grsecurity.net>
Date:   Wed Sep 19 20:57:26 2012 -0400

    allow kernel pointers to be printed in /proc/vmallocinfo, a 0600 entry

 mm/vmalloc.c |    2 +-
 1 files changed, 1 insertions(+), 1 deletions(-)

commit 1ca8b4bf54b3fa594c5fc502902c51453ef5a28b
Merge: 38d0916 e425144
Author: Brad Spengler <spender@grsecurity.net>
Date:   Wed Sep 19 19:58:45 2012 -0400

    Merge branch 'pax-stable' into grsec-stable

commit e425144244f00b460449983d680fbf3c1e8077c5
Author: Brad Spengler <spender@grsecurity.net>
Date:   Wed Sep 19 19:58:08 2012 -0400

    Update to pax-linux-2.6.32.59-test198.patch:
    - fixed several copy_*_user_overflow induced warnings
    - fixed fallout of struct apic constification in xen

 arch/x86/kernel/tls.c                      |    2 +-
 arch/x86/xen/enlighten.c                   |   12 ++++++------
 drivers/net/wireless/iwlwifi/iwl-debugfs.c |   10 +++++-----
 kernel/capability.c                        |    2 +-
 net/bluetooth/hci_sock.c                   |    2 +-
 net/bluetooth/rfcomm/sock.c                |    2 +-
 net/compat.c                               |    2 +-
 net/ipv4/netfilter/arp_tables.c            |    6 +++---
 net/ipv4/netfilter/ip_tables.c             |    6 +++---
 net/ipv6/netfilter/ip6_tables.c            |    6 +++---
 net/packet/af_packet.c                     |    4 ++--
 security/Kconfig                           |    3 ---
 12 files changed, 27 insertions(+), 30 deletions(-)

commit 38d0916c8f0cc53205f43a8ec56133f5a65ff24d
Author: Brad Spengler <spender@grsecurity.net>
Date:   Wed Sep 19 18:05:27 2012 -0400

    From: Mathias Krause <minipli@googlemail.com>
    To: "David S. Miller" <davem@davemloft.net>
    Cc: Steffen Klassert <steffen.klassert@secunet.com>, netdev@vger.kernel.org,
            linux-kernel@vger.kernel.org,
            Mathias Krause <minipli@googlemail.com>
    Subject: [PATCH 3/6] xfrm_user: fix info leak in copy_to_user_policy()
    Date: Wed, 19 Sep 2012 23:33:40 +0200

    The memory reserved to dump the xfrm policy includes multiple padding
    bytes added by the compiler for alignment (padding bytes in struct
    xfrm_selector and struct xfrm_userpolicy_info). Add an explicit
    memset(0) before filling the buffer to avoid the heap info leak.


 net/xfrm/xfrm_user.c |    1 +
 1 files changed, 1 insertions(+), 0 deletions(-)

Signature-tree: 6f602816800dd815aecb49b656cbf26831afbf2b

+++ drivers/net/wireless/iwlwifi/iwl-debugfs.c
@@ -166,7 +166,7 @@ static ssize_t iwl_dbgfs_tx_statistics_write(struct file *file,
 	struct iwl_priv *priv = file->private_data;
 	u32 clear_flag;
 	char buf[8];
-	int buf_size;
+	size_t buf_size;
 
 	memset(buf, 0, sizeof(buf));
 	buf_size = min(count, sizeof(buf) -  1);
@@ -285,7 +285,7 @@ static ssize_t iwl_dbgfs_sram_write(struct file *file,
 {
 	struct iwl_priv *priv = file->private_data;
 	char buf[64];
-	int buf_size;
+	size_t buf_size;
 	u32 offset, len;
 
 	memset(buf, 0, sizeof(buf));
@@ -743,7 +743,7 @@ static ssize_t iwl_dbgfs_disable_ht40_write(struct file *file,
 {
 	struct iwl_priv *priv = file->private_data;
 	char buf[8];
-	int buf_size;
+	size_t buf_size;
 	int ht40;
 
 	memset(buf, 0, sizeof(buf));
@@ -786,7 +786,7 @@ static ssize_t iwl_dbgfs_sleep_level_override_write(struct file *file,
 {
 	struct iwl_priv *priv = file->private_data;
 	char buf[8];
-	int buf_size;
+	size_t buf_size;
 	int value;
 
 	memset(buf, 0, sizeof(buf));
@@ -954,7 +954,7 @@ static ssize_t iwl_dbgfs_traffic_log_write(struct file *file,
 {
 	struct iwl_priv *priv = file->private_data;
 	char buf[8];
-	int buf_size;
+	size_t buf_size;
 	int traffic_log;
 
 	memset(buf, 0, sizeof(buf));
